gitops-controller/requirements.txt
# gitops-controller/requirements.txt
flask==3.0.0
requests==2.31.0
numpy==1.26.2
scikit-learn==1.3.2
pyyaml==6.0.1
gitpython==3.1.40